refactor: extract AiSettings panel from SettingsDialog (#672)
Fifth per-tab panel split, and the last big one. The ai tab — default model +
Anthropic API key + per-tool model overrides — moves into AiSettings.svelte.
SettingsDialog 1363 → 1162.

Unlike the other panels, the ai settings persist on the dialog's Done handler
(so the key + model + overrides apply together), not on each edit. So this one
is presentational, not self-contained: the dialog keeps the state + handleDone
and binds it in via $bindable props (model / apiKeyInput / clearApiKey /
toolModelOverrides), with apiKeyStatus read-only. The save-on-Done behavior is
unchanged. The dialog sheds its now-unused MODEL_OPTIONS / modelLabel /
getAllToolInfos / allTools / setToolOverride and the .tool-models /
.api-key-status / .link-btn CSS.

Tests (5): components/AiSettings.test.ts mocks getAllToolInfos and pins the API-
key status states, the Clear-saved-key → cleared (input disabled) → Cancel-clear
round trip, the tool-override table (name + preference + per-tool select), the
no-tools empty state, and an override select reflecting the chosen model.

Verified: pnpm lint (0 errors), full suite (3010 passed, +5), electron-forge
build + boot smoke.
